prise en compte des idx manquant

+++ script/generate_mdtm_from_xgtf.py
@@ -1,4 +1,4 @@
-import glob, sys
+import glob, sys, os
 
 path_xgtf = sys.argv[1]
 path_idx = sys.argv[2]
@@ -34,43 +34,46 @@ for f_xgtf in glob.glob(path_xgtf+'/*.xgtf'):
 
     video = f_xgtf.split('/')[-1].split('.')[0]
 
-    dic_idx = sync_idx(path_idx+video+'.MPG.idx')
-    fout = open(path_out+video+'.mdtm', 'w')
+    if os.path.isfile(path_idx+video+'.MPG.idx'):
 
-    data=False
-    frame=0
-    doc = []
-    current_line=''
-    for line in open(f_xgtf):
-        if data:
-            current_line+=line.replace('\n','')            
-            if '</object>' in line or '</file>' in line:
-                doc.append(current_line)
-                current_line=''            
-        if '<data>' in line:
-            data=True      
+        dic_idx = sync_idx(path_idx+video+'.MPG.idx')
+        fout = open(path_out+video+'.mdtm', 'w')
 
-    start_tete=False
-    l_frame=[]
-    for line in doc:
-        line = line.replace('&lt;', '<')
-        if 'name="PERSONNE"' in line:
-            l = line.split('<attribute')  
-            if 'framespan="' in l[0]:
-                name=''
-                startframe=''
-                endframe='' 
-                for a in l:
-                    if 'name="NOM">' in a:
-                        name= a.split('svalue value="')[1].split('"/>')[0]                    
-                    if 'name="STARTFRAME">' in a:
-                        startframe= a.split('dvalue value="')[1].split('"/>')[0]
-                    if 'name="ENDFRAME">' in a:
-                        endframe= a.split('dvalue value="')[1].split('"/>')[0]
-                if name!='' and startframe!='' and endframe!='' : 
-                    start = dic_idx[int(startframe)]
-                    duration = dic_idx[int(endframe)] - start
-                    fout.write(video+' 1 '+str(start)+' '+str(duration)+' head NA NA '+name+'\n')
-    fout.close()
+        data=False
+        frame=0
+        doc = []
+        current_line=''
+        for line in open(f_xgtf):
+            if data:
+                current_line+=line.replace('\n','')            
+                if '</object>' in line or '</file>' in line:
+                    doc.append(current_line)
+                    current_line=''            
+            if '<data>' in line:
+                data=True      
+
+        start_tete=False
+        l_frame=[]
+        for line in doc:
+            line = line.replace('&lt;', '<')
+            if 'name="PERSONNE"' in line:
+                l = line.split('<attribute')  
+                if 'framespan="' in l[0]:
+                    name=''
+                    startframe=''
+                    endframe='' 
+                    for a in l:
+                        if 'name="NOM">' in a:
+                            name= a.split('svalue value="')[1].split('"/>')[0]                    
+                        if 'name="STARTFRAME">' in a:
+                            startframe= a.split('dvalue value="')[1].split('"/>')[0]
+                        if 'name="ENDFRAME">' in a:
+                            endframe= a.split('dvalue value="')[1].split('"/>')[0]
+                    if name!='' and startframe!='' and endframe!='' : 
+                        if int(startframe) in dic_idx and int(endframe) in dic_idx:
+                            start = dic_idx[int(startframe)]
+                            duration = dic_idx[int(endframe)] - start
+                            fout.write(video+' 1 '+str(start)+' '+str(duration)+' head NA NA '+name+'\n')
+        fout.close()
